services/ocr/requirements.txt
Pillow==2.3.0
pyocr==0.4.1
